requirements.txt
# Telnyx Voice API
telnyx>=2.1.0

# Deepgram SDK for STT and TTS
deepgram-sdk>=3.0.0

# Web framework for WebSocket server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Async HTTP client
aiohttp>=3.9.0

# Environment management
python-dotenv>=1.0.0

# Audio processing
audioop-lts>=0.2.1

# SSL certificates
certifi>=2024.0.0

